requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-groq===0.1.9
python-dotenv==1.0.0
streamlit==1.28.1
PyPDF2==3.0.1
sentence-transformers
chromadb==0.4.17
protobuf==3.20
pysqlite3-binary
